fix zshprofile

+++ zsh/.zshenv
@@ -1,2 +1,5 @@
-ln -s $HOME/.Xauthority /run/user/1000/Xauthority
+if [ ! -f "$HOME/.Xauthority" ]; then
+  ln -s $HOME/.Xauthority /run/user/1000/Xauthority
+fi
+
 export ZDOTDIR=$HOME/.config/zsh
